+++ web/app/themes/brocooly/config/app.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Application main configuration file
+ * Application configuration file
  *
  * @package Brocooly
  * @subpackage Brocket
@@ -23,5 +23,67 @@ return [
 	 * @var string
 	 */
 	'locale' => get_locale(),
+
+	/**
+	 * --------------------------------------------------------------------------
+	 * Assets configuration
+	 * --------------------------------------------------------------------------
+	 *
+	 * Define rules for assets, compiled by webpack. Regex defines which asset
+	 * is a script and which one is a style within `mix-manifest.json`
+	 * Matched values within styles array will be loaded via `wp_enqueue_style`
+	 * Scripts will be loaded with `wp_enqueue_script()` in a footer
+	 *
+	 * Queue is an array of arrays of assets to be load conditionally
+	 *
+	 * @since 1.7.2
+	 * @var array
+	 */
+	'assets' => [
+
+		/**
+		 * --------------------------------------------------------------------------
+		 * Assets configuration
+		 * --------------------------------------------------------------------------
+		 *
+		 * Load assets or not by default. Set to `false` of you wish register
+		 * styles and scripts by your own
+		 *
+		 * @var array
+		 */
+		'autoload' => true,
+
+		'styles'   => [
+			'regex' => '/(\/css\/)[\w]+\.css$/',
+
+			/**
+			 * @example
+			 * ```
+			 * 'queue' => [
+			 * 		[
+			 * 			'key' => '/css/app.css' // same as within `mix-manifest.json`
+			 * 			'condition' => 'is_front_page' // condition to load style
+			 * 		],
+			 * ],
+			 *
+			 * Extra keys are `version`, `media` and `deps` - same as for `wp_enqueue_style()`
+			 * ```
+			 */
+			'queue' => [],
+		],
+
+		'scripts'  => [
+			'regex' => '/(\/js\/)[\w]+\.js$/',
+
+			/**
+			 * @example
+			 * ```
+			 * Same as for styles but instead of `media` key there is `inFooter`
+			 * ```
+			 */
+			'queue' => [],
+		],
+
+	],
 
 ];